MediaAnalyzerFactory
