+++ IntroSkipper/ScheduledTasks/BaseItemAnalyzerTask.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
@@ -18,26 +19,36 @@ namespace IntroSkipper.ScheduledTasks;
 /// <summary>
 /// Common code shared by all media item analyzer tasks.
 /// </summary>
-/// <remarks>
-/// Initializes a new instance of the <see cref="BaseItemAnalyzerTask"/> class.
-/// </remarks>
-/// <param name="logger">Task logger.</param>
-/// <param name="loggerFactory">Logger factory.</param>
-/// <param name="libraryManager">Library manager.</param>
-/// <param name="mediaSegmentUpdateManager">Media segment update manager.</param>
-public class BaseItemAnalyzerTask(
-    ILogger logger,
-    ILoggerFactory loggerFactory,
-    ILibraryManager libraryManager,
-    MediaSegmentUpdateManager mediaSegmentUpdateManager)
+public class BaseItemAnalyzerTask
 {
-    private readonly ILogger _logger = logger;
-    private readonly ILoggerFactory _loggerFactory = loggerFactory;
-    private readonly ILibraryManager _libraryManager = libraryManager;
-    private readonly MediaSegmentUpdateManager _mediaSegmentUpdateManager = mediaSegmentUpdateManager;
+    private readonly ILogger _logger;
+    private readonly ILoggerFactory _loggerFactory;
+    private readonly ILibraryManager _libraryManager;
+    private readonly MediaSegmentUpdateManager _mediaSegmentUpdateManager;
     private readonly PluginConfiguration _config = Plugin.Instance?.Configuration ?? new PluginConfiguration();
     private readonly bool _ffmpegValid = FFmpegWrapper.CheckFFmpegVersion();
-    private Dictionary<AnalysisMode, List<IMediaFileAnalyzer>> _analyzers = [];
+    private readonly MediaAnalyzerFactory _analyzerFactory;
+    private Dictionary<AnalysisMode, Collection<IMediaFileAnalyzer>> _analyzers = [];
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="BaseItemAnalyzerTask"/> class.
+    /// </summary>
+    /// <param name="logger">Task logger.</param>
+    /// <param name="loggerFactory">Logger factory.</param>
+    /// <param name="libraryManager">Library manager.</param>
+    /// <param name="mediaSegmentUpdateManager">Media segment update manager.</param>
+    public BaseItemAnalyzerTask(
+        ILogger logger,
+        ILoggerFactory loggerFactory,
+        ILibraryManager libraryManager,
+        MediaSegmentUpdateManager mediaSegmentUpdateManager)
+    {
+        _logger = logger;
+        _loggerFactory = loggerFactory;
+        _libraryManager = libraryManager;
+        _mediaSegmentUpdateManager = mediaSegmentUpdateManager;
+        _analyzerFactory = new MediaAnalyzerFactory(loggerFactory, _config, _ffmpegValid);
+    }
 
     /// <summary>
     /// Analyze all media items on the server.
@@ -71,7 +82,7 @@ public class BaseItemAnalyzerTask(
                 "If Jellyfin is running in a container, upgrade to version 10.10.0 or newer.");
         }
 
-        _analyzers = GetAnalyzers();
+        _analyzers = (Dictionary<AnalysisMode, Collection<IMediaFileAnalyzer>>)_analyzerFactory.GetAnalyzers();
 
         var modes = _analyzers.Keys;
 
@@ -178,7 +189,7 @@ public class BaseItemAnalyzerTask(
             first.SeriesName,
             first.SeasonNumber);
 
-        var analyzers = GetFilteredAnalyzers(first, mode, _analyzers[mode]);
+        var analyzers = _analyzerFactory.GetFilteredAnalyzers(first, mode, _analyzers[mode]);
 
         // Use each analyzer to find skippable ranges in all media files, removing successfully
         // analyzed items from the queue.
@@ -192,101 +203,5 @@ public class BaseItemAnalyzerTask(
         await Plugin.Instance!.SetEpisodeIdsAsync(first.SeasonId, mode, items.Select(i => i.EpisodeId)).ConfigureAwait(false);
 
         return totalItems - items.Count(e => e.GetAnalyzed(mode) != EpisodeState.Analyzed);
-    }
-
-    private Dictionary<AnalysisMode, List<IMediaFileAnalyzer>> GetAnalyzers()
-    {
-        var analyzers = new Dictionary<AnalysisMode, List<IMediaFileAnalyzer>>();
-
-        // Create analyzer lists
-        var introAnalyzers = ParseAnalyzers(_config.IntroAnalyzerOrderList);
-        var creditsAnalyzers = ParseAnalyzers(_config.CreditsAnalyzerOrderList, includeBlackFrame: true);
-        var chapterAnalyzer = new List<IMediaFileAnalyzer> { new ChapterAnalyzer(_loggerFactory.CreateLogger<ChapterAnalyzer>()) };
-
-        // Populate the dictionary
-        if (introAnalyzers.Count > 0)
-        {
-            analyzers.Add(AnalysisMode.Introduction, introAnalyzers);
-        }
-
-        if (creditsAnalyzers.Count > 0)
-        {
-            analyzers.Add(AnalysisMode.Credits, creditsAnalyzers);
-        }
-
-        if (_config.ScanRecap)
-        {
-            analyzers.Add(AnalysisMode.Recap, chapterAnalyzer);
-        }
-
-        if (_config.ScanPreview)
-        {
-            analyzers.Add(AnalysisMode.Preview, chapterAnalyzer);
-        }
-
-        return analyzers;
-    }
-
-    private List<IMediaFileAnalyzer> ParseAnalyzers(string? configString, bool includeBlackFrame = false)
-    {
-        var result = new List<IMediaFileAnalyzer>();
-        if (string.IsNullOrEmpty(configString))
-        {
-            return result;
-        }
-
-        var analyzerItems = configString.Split(',')
-            .Select(item =>
-            {
-                var parts = item.Split(':', 2);
-                return (Name: parts[0].Trim(), Enabled: bool.Parse(parts[1]));
-            })
-            .Where(analyzer => analyzer.Enabled)
-            .ToList();
-
-        foreach (var (name, _) in analyzerItems)
-        {
-            IMediaFileAnalyzer? instance = name switch
-            {
-                "Chapter" => new ChapterAnalyzer(_loggerFactory.CreateLogger<ChapterAnalyzer>()),
-                "Chromaprint" when _ffmpegValid => new ChromaprintAnalyzer(_loggerFactory.CreateLogger<ChromaprintAnalyzer>()),
-                "BlackFrame" when includeBlackFrame => new BlackFrameAnalyzer(_loggerFactory.CreateLogger<BlackFrameAnalyzer>()),
-                _ => null
-            };
-
-            if (instance is not null)
-            {
-                result.Add(instance);
-            }
-        }
-
-        return result;
-    }
-
-    private List<IMediaFileAnalyzer> GetFilteredAnalyzers(QueuedEpisode first, AnalysisMode mode, List<IMediaFileAnalyzer> analyzers)
-    {
-        List<IMediaFileAnalyzer> result = Plugin.Instance!.GetAnalyzerAction(first.SeasonId, mode) switch
-        {
-            AnalyzerAction.Chapter => [.. analyzers.OfType<ChapterAnalyzer>()],
-            AnalyzerAction.Chromaprint => [.. analyzers.OfType<ChromaprintAnalyzer>()],
-            AnalyzerAction.BlackFrame => [.. analyzers.OfType<BlackFrameAnalyzer>()],
-            AnalyzerAction.None => [],
-            _ when first.IsMovie => [.. analyzers.Where(a => a is not ChromaprintAnalyzer)],
-            _ => [.. analyzers]
-        };
-
-        if (result.Count > 1 && first.IsAnime && _config.AnimeChromaprint && mode == AnalysisMode.Credits)
-        {
-            int chromaprintIndex = result.FindIndex(a => a is ChromaprintAnalyzer);
-            int blackFrameIndex = result.FindIndex(a => a is BlackFrameAnalyzer);
-
-            // Swap the analyzers if needed
-            if (chromaprintIndex != -1 && blackFrameIndex != -1 && chromaprintIndex > blackFrameIndex)
-            {
-                (result[blackFrameIndex], result[chromaprintIndex]) = (result[chromaprintIndex], result[blackFrameIndex]);
-            }
-        }
-
-        return result;
     }
 }
